Assembled enkf_state functions accessing node_hash.
